pyproject.toml
[tool.poetry]
name = "hackabot"
version = "0.1.0"
description = ""
authors = ["IDenra <denrazv@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
granula = "^2018.2.8"
requests = "^2.24.0"
PyTelegramBotAPI = "^3.7.3"
numpy = "^1.19.2"
google = "^3.0.0"
google-cloud = "^0.34.0"
tqdm = "^4.48.2"
googleapis-common-protos = "^1.52.0"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
